refactor: weaken types, input streaming adjustments & missing tests

+++ src/test/java/com/gotreaux/aoc/puzzles/year2015/day5/NiceStringPuzzleTest.java
@@ -18,7 +18,7 @@ class NiceStringPuzzleTest {
 
         NiceStringPuzzle puzzle = new NiceStringPuzzle(inputProvider);
 
-        PuzzleOutput<Integer, Integer> output = puzzle.solve();
+        PuzzleOutput<Long, Long> output = puzzle.solve();
 
         assertEquals(expectedCount, output.partOne());
     }
@@ -30,7 +30,7 @@ class NiceStringPuzzleTest {
 
         NiceStringPuzzle puzzle = new NiceStringPuzzle(inputProvider);
 
-        PuzzleOutput<Integer, Integer> output = puzzle.solve();
+        PuzzleOutput<Long, Long> output = puzzle.solve();
 
         assertEquals(expectedCount, output.partTwo());
     }
